Move non click files to bris_adapt/process.
